% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclid.r
\name{euclid}
\alias{euclid}
\title{Euclidean distance}
\usage{
euclid(a, b)
}
\arguments{
\item{a}{Numeric vector.}

\item{b}{Numeric vector of same length as \code{a}.}
}
\value{
Numeric.
}
\description{
Euclidian distance in any dimension
}
\examples{
euclid(0, 5)
euclid(c(0, 0), c(1, 1))
euclid(c(0, 0, 0), c(1, 1, 1))
}
