% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsd.r
\name{rmsd}
\alias{rmsd}
\title{Root-mean-square deviation}
\usage{
rmsd(x1, x2, w = NULL, na.rm = FALSE)
}
\arguments{
\item{x1}{Numeric vector, matrix, or data frame.}

\item{x2}{Numeric vector the same length as \code{x1}, or a matrix or data frame the same dimensions as \code{x1}.}

\item{w}{Weights. If \code{x1} and \code{x2} are vectors, this is a numeric vector the same length as \code{x1} or \code{x2}. If \code{x1} and \code{x2} are matrices or data frames then this is either a matrix or data frame with the same dimensions as \code{x1} and \code{x2}. The default value of \code{NULL} assigns each pair of values in \code{x1} and \code{x2} equal weight.}

\item{na.rm}{Logical, if \code{TRUE} then remove any elements in \code{x1} \emph{and} \code{x2} where either \code{x1} or \code{x2} is \code{NA}. Default is \code{FALSE}, in which case any \code{NA} returns \code{NA}.}
}
\value{
Numeric.
}
\description{
Calculate the root-mean-square deviation (\code{sqrt(mean((x1 - x2)^2))}). If non-constant weights \code{w} are supplied, then the calculation is \code{sqrt(sum(w * (x1 - x2)^2) / sum(w))}.
}
\examples{
set.seed(123)
# numeric vectors
x1 <- 1:20
x2 <- 1:20 + rnorm(20)
rmsd(x1, x2)
x1[1] <- NA
rmsd(x1, x2)
rmsd(x1, x2, na.rm=TRUE)

# matrices
x1 <- matrix(runif(20), ncol=5)
x2 <- matrix(rnorm(20), ncol=5)
rmsd(x1, x2)
x1[1, 1] <- NA
rmsd(x1, x2)
rmsd(x1, x2, na.rm=TRUE)
}
