% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsd.r
\name{rmsd}
\alias{rmsd}
\title{Root-mean-square deviation}
\usage{
rmsd(x1, x2, na.rm = FALSE)
}
\arguments{
\item{x1}{Numeric vector.}

\item{x2}{Numeric vector same length as \code{x1}.}

\item{na.rm}{Logical, if \code{TRUE} then remove any elements in \code{x1} \emph{and} \code{x2} where either \code{x1} or \code{x2} is \code{NA}. Default is \code{FALSE}, in which case any \code{NA} returns \code{NA}.}
}
\value{
Numeric.
}
\description{
Calculate the root-mean-square deviation (\code{sqrt(mean((x1 - x2)^2))}).
}
\examples{
set.seed(123)
x1 <- 1:20
x2 <- 1:20 + rnorm(20)
rmsd(x1, x2)
x1[1] <- NA
rmsd(x1, x2)
rmsd(x1, x2, na.rm=TRUE)
}
